That’s a big commit!
Add custom font for icons, made with IcoMoon.

Better CSS

Now, a document’s details can be viewed when clicking on the info
button, can delete the document when clicking on the bin icon when the
user is connected only.

Convert a lot of option API scripts to setup API

Some pages and actions only show up when the user is connected

Change some fetch requests to axios requests, may change to vueuse’s
useFetch though